av1.requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7